Bring code to 2024 standards
